refactor!: Make routes public by default and rename routes
- Make routes public by default and specify protected routes by
  appending to `authRoutes`
- Rename `/join-us` -> `/join`
- Rename `/sign-in` -> `/signin`
- Rename `utils` to `lib`

Resolves #36
