+++ src/app/(account)/join/ProgressBar.tsx
