+++ src/app/(account)/join/page.tsx
