+++ src/app/(account)/join/steps/StepTwo.tsx
