+++ src/app/(account)/join/store.ts
