+++ src/lib/square.ts
@@ -1,0 +1,8 @@
+import { Client, Environment } from 'square';
+import { env } from '@/env.mjs';
+
+export const squareClient = new Client({
+    accessToken: env.SQUARE_ACCESS_TOKEN,
+    environment:
+        process.env.NODE_ENV === 'production' ? Environment.Production : Environment.Sandbox,
+});
